Backend/requirements.txt
amqp==5.3.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.9.2
attrs==25.3.0
billiard==4.2.2
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
cron_descriptor==2.0.6
cryptography==46.0.1
defusedxml==0.7.1
Django==5.2.6
django-allauth==65.11.2
django-anymail==13.1
django-celery-beat==2.8.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-countries==7.6.1
django-debug-toolbar==6.0.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-phonenumber-field==8.1.0
django-redis==6.0.0
django-silk==5.4.3
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
dotenv==0.9.9
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.9.1
et_xmlfile==2.0.0
factory_boy==3.3.3
Faker==37.8.0
gprof2dot==2025.4.14
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
phonenumbers==9.0.14
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python3-openid==3.2.0
PyYAML==6.0.3
redis==6.4.0
referencing==0.36.2
reportlab==4.4.4
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
sentry-sdk==2.39.0
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
